requirements.txt
beautifulsoup4==4.8.1
Click==7.0
Markdown==3.1.1
pydub==0.23.1
google-cloud-texttospeech==0.5.0
